+++ src/components/User/Users/components/UserDetails/UserDetails.vue
@@ -199,7 +199,7 @@ export default {
       isShownExternal: false,
       user: {},
       requests: [],
-      verifiedRequest: {},
+      verifiedRequest: new ChangeRoleRequest({}),
     }
   },
 
@@ -311,7 +311,7 @@ export default {
       setTimeout(async () => {
         await this.getUser()
         this.$emit(EVENTS.reviewed)
-      }, 1500)
+      }, 3000)
     },
   },
 }